refactor: colocate pass 3 api route tests
