+++ app/api/boss-quests/[id]/complete/route.achievement-integration.test.ts
