+++ app/api/boss-quests/[id]/complete/route.helpers.test.ts
