+++ app/api/quest-instances/[id]/assign/route.test.ts
@@ -6,7 +6,7 @@ import {
   params,
   singleResult,
   makeSetupAuth,
-} from "./quest-instance-helpers";
+} from "../quest-instance-helpers";
 
 const mockSupabase = {
   auth: { getUser: jest.fn() },
