+++ app/api/quest-instances/[id]/quest-instance-helpers.ts
