+++ app/api/quest-instances/[id]/release/route.test.ts
@@ -8,7 +8,7 @@ import {
   params,
   singleResult,
   makeSetupAuth,
-} from "./quest-instance-helpers";
+} from "../quest-instance-helpers";
 
 const mockSupabase = {
   auth: { getUser: jest.fn() },
