+++ app/api/quest-instances/[id]/route.delete.test.ts
