+++ app/api/quests/[id]/route.error-handling.test.ts
